Мелкие исправления

+++ Chillinator 2/Components/Disk.swift
@@ -106,7 +106,7 @@ import Kingfisher
     
     
     /// Установка позиции головки проигрывателя
-    func setPlayHeadPosition(relativeTime: Double) {
+    func setPlayHeadPosition(relativeTime: Float) {
         let angle = -CGFloat(pow(relativeTime, 0.8)) * CGFloat.pi / 12
         let options: UIView.AnimationOptions = relativeTime.isZero ? .curveEaseInOut : .curveLinear
         UIView.animate(withDuration: 1, delay: 0, options: options, animations: { [weak self] in
